+++ go/sequence/sequence.go
@@ -1,7 +1,7 @@
 package sequence
 
 type Sequencer interface {
-	NextFileId(count int) (uint64, int)
+	NextFileId(count uint64) (uint64, uint64)
 	SetMax(uint64)
 	Peek() uint64
 }
